MAJOR LOGIC CHANGE - READ DESCRIPTION
Previous to this commit, the process for checking keys was as follows:

Check global keys for a match. Failing that:
Check room-global keys for a match. Failing that:
Check roomstate-specific keys for a match. Failing that:
print I don't understand that command.

THE NEW SYSTEM IS REVERSED. IT IS NOW:

Check roomstate-specific keys for a match. Failing that:
Check room-global keys for a match. Failing that:
Check global keys for a match. Failing that:
print I don't understand that command.